[Samantha844] Trabajo terminado

+++ routes/web.php
@@ -24,9 +24,33 @@ Route::get('/empleado/index', 'EmpleadoController@index')->name('empleado.index'
 Route::get('/empleado/create', 'EmpleadoController@create')->name('empleado.create');
 Route::post('/empleado/store', 'EmpleadoController@store')->name('empleado.store');
 
-Route::delete('/empleado/{empleado}', 'EmpleadoController@destroy')->name('empleado.destroy');
+Route::get('/empleado/{empleado}/show', 'EmpleadoController@show')->name('empleado.show');
+
+Route::get('/empleado/create','EmpleadoController@create')->name('empleado.create')->middleware("editDeleteInsert.admin");;
+
+Route::delete('/empleado/{empleado}','EmpleadoController@destroy')->name('empleado.destroy')->middleware("editDeleteInsert.admin");;
+
+Route::get('/empleado/{empleado}/edit','EmpleadoController@edit')->name('empleado.edit')->middleware("editDeleteInsert.admin");
+
+Route::put('/empleado/{empleado}','EmpleadoController@update')->name('empleado.update');
 
 
 Route::post('/changeLang', 'HomeController@changeLang')->name('changeLang');
 
 Route::get('/changeLangGet/{locale_id}', 'HomeController@changeLangGet')->name('changeLangGet');
+Auth::routes();
+
+
+
+Route::get('datoContacto','DatoContactoController@index')->name('datoContacto.index');
+Route::get('datoContacto/{empleado}/create','DatoContactoController@create')->name('datoContacto.create');
+Route::post('datoContacto','DatoContactoController@store')->name('datoContacto.store');
+
+Route::get('datoContacto/{datoContacto}/show','DatoContactoController@show')->name('datoContacto.show');
+
+Route::get('datoContacto/{datoContacto}/edit','DatoContactoController@edit')->name('datoContacto.edit');
+Route::put('datoContacto/{datoContacto}','DatoContactoController@update')->name('datoContacto.update');
+
+Route::delete('datoContacto/{datoContacto}/{empleadoId}','DatoContactoController@destroy')->name('datoContacto.destroy');
+
+Route::get('/home', 'HomeController@index')->name('home');
